Organize mocks and service injection
- Mock classes are now in a separate folder with a separate namespace.
- Mock objects are now injected in unit/e2e tests when necessary.
- Conditional DI of ISchoolService in Startup is removed.
