+++ src/Services/ISchoolService.cs
@@ -1,0 +1,4 @@
+public interface ISchoolService
+{
+	System.Collections.Generic.IEnumerable<KonSchool.Models.School> GetSchools();
+}
