+++ tests/ApiControllerTests/SchoolsController.test.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using KonSchool.Models;
 using KonSchool.ApiControllers;
-using KonSchool.Services;
+using KonSchool.Tests.Mocks;
 
 namespace KonSchool.Tests.ApiControllerTests
 {
@@ -13,7 +13,7 @@ namespace KonSchool.Tests.ApiControllerTests
         [Fact]
         public void Can_Get_AllSchools()
         {
-            var ctrlr = new SchoolsController(new SchoolServiceMock());
+            var ctrlr = new SchoolsController(new MockSchoolService());
             var response = ctrlr.Get();
             Assert.IsType<OkObjectResult>(response.Result);
             var AllSchools = (response.Result as OkObjectResult).Value as IEnumerable<School>;
@@ -23,17 +23,17 @@ namespace KonSchool.Tests.ApiControllerTests
         [Fact]
         public void Can_Get_OneSchool_EIINwise()
         {
-            var ctrlr = new SchoolsController(new SchoolServiceMock());
+            var ctrlr = new SchoolsController(new MockSchoolService());
             var response = ctrlr.GetOneSchool("1212");
             Assert.IsType<OkObjectResult>(response.Result);
             var OneSchool = (response.Result as OkObjectResult).Value as School;
             Assert.Equal("Comilla", OneSchool.Division);
-        }        
+        }
 
         [Fact]
         public void Can_Get_AllSchools_DivWise()
         {
-            var ctrlr = new SchoolsController(new SchoolServiceMock());
+            var ctrlr = new SchoolsController(new MockSchoolService());
             var response = ctrlr.GetSchoolsByDivision("Dhaka");
             Assert.IsType<OkObjectResult>(response.Result);
             var resultValue = (response.Result as OkObjectResult).Value;
@@ -41,12 +41,12 @@ namespace KonSchool.Tests.ApiControllerTests
             Assert.Equal(4, SchoolsInDhkDiv.Count);
             foreach (var school in SchoolsInDhkDiv)
                 Assert.Equal("Dhaka", school.Division);
-        }       
+        }
 
         [Fact]
         public void Can_Get_AllSchools_DisWise()
         {
-            var ctrlr = new SchoolsController(new SchoolServiceMock());
+            var ctrlr = new SchoolsController(new MockSchoolService());
             var response = ctrlr.GetSchoolsByDistrict("Faridpur");
             Assert.IsType<OkObjectResult>(response.Result);
             var OneSchoolInFaridpur = (response.Result as OkObjectResult).Value as IEnumerable<School>;
@@ -57,16 +57,16 @@ namespace KonSchool.Tests.ApiControllerTests
         [Fact]
         public void Cannot_Get_NonexistentSchools()
         {
-            var ctrlr = new SchoolsController(new SchoolServiceMock());
+            var ctrlr = new SchoolsController(new MockSchoolService());
 
             // EIIN wise
             var response_EIIN = ctrlr.GetOneSchool("2121");
             Assert.IsType<NotFoundResult>(response_EIIN.Result);
-            
+
             // Div wise
             var response_Division = ctrlr.GetSchoolsByDivision("NewYork");
             Assert.IsType<NotFoundResult>(response_Division.Result);
-            
+
             // Dist wise
             var response_District = ctrlr.GetSchoolsByDistrict("Gotham");
             Assert.IsType<NotFoundResult>(response_District.Result);
