+++ tests/EndToEndTests/PageLoadTests.cs
@@ -1,6 +1,9 @@
 using Xunit;
 using Microsoft.AspNetCore.Mvc.Testing;
 using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection;
+using KonSchool.Services;
+using KonSchool.Tests.Mocks;
 
 namespace KonSchool.Tests.EndToEndTests
 {
@@ -25,7 +28,10 @@ namespace KonSchool.Tests.EndToEndTests
         public async Task Get_EndpointsReturnSuccessAndCorrectContentType(string url)
         {
             // Arrange
-            var client = factory.CreateClient();
+            var client = factory.WithWebHostBuilder(builder =>
+				builder.ConfigureServices(services => services.AddSingleton<ISchoolService, MockSchoolService>())
+			)
+			.CreateClient();
 
             // Act
             var response = await client.GetAsync(url);
