+++ tests/ModelTests/Query.test.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using KonSchool.Models;
-using KonSchool.Services;
+using KonSchool.Tests.Mocks;
 using Xunit;
 
 namespace KonSchool.Tests.ModelTests
@@ -10,7 +10,7 @@ namespace KonSchool.Tests.ModelTests
         [Fact]
         public void CtorTest()
         {
-            Query query = new Query(new SchoolServiceMock())
+            Query query = new Query(new MockSchoolService())
             {
                 Class = 10,
                 Social = 10.0,
@@ -34,7 +34,7 @@ namespace KonSchool.Tests.ModelTests
         [Fact]
         public void Can_SetLocation()
         {
-            Query query = new Query(new SchoolServiceMock());
+            Query query = new Query(new MockSchoolService());
             query.SetLocation("Dhaka", "Dhaka", "Rampura", "Ward no. 1");
             Assert.Equal("Dhaka", query.Division);
             Assert.Equal("Dhaka", query.District);
@@ -45,7 +45,7 @@ namespace KonSchool.Tests.ModelTests
         [Fact]
         public void Can_CheckEligibility_Female()
         {
-            Query query = new Query(new SchoolServiceMock())
+            Query query = new Query(new MockSchoolService())
             {
                 Class = 10,
                 Social = 10.0,
@@ -77,7 +77,7 @@ namespace KonSchool.Tests.ModelTests
         [Fact]
         public void Can_CheckEligibility_Male()
         {
-            Query query = new Query(new SchoolServiceMock())
+            Query query = new Query(new MockSchoolService())
             {
                 Class = 7,
                 Social = 10.0,
@@ -95,7 +95,7 @@ namespace KonSchool.Tests.ModelTests
             };
 
             query.SetValues();
-            
+
             Assert.Equal(4, query.Alternatives.Count);
 
             query.Division = "Dhaka";
