+++ tests/PageModelTests/OutputsModel.test.cs
@@ -1,7 +1,7 @@
 using Xunit;
 using KonSchool.Models;
 using KonSchool.Pages;
-using KonSchool.Services;
+using KonSchool.Tests.Mocks;
 using Microsoft.AspNetCore.Http;
 using Moq;
 
@@ -40,7 +40,7 @@ namespace KonSchool.Tests.PageModelTests
 		[Fact]
         public void Can_OnGet()
         {
-			var outputsPage = new OutputsModel(new SchoolServiceMock(), httpCtxAccessor.Object);
+			var outputsPage = new OutputsModel(new MockSchoolService(), httpCtxAccessor.Object);
 			outputsPage.OnGet();
 			Assert.Equal(6, outputsPage.Criteria.Length);
 		}
@@ -48,7 +48,7 @@ namespace KonSchool.Tests.PageModelTests
         [Fact]
         public void Can_ConvertNumbers_ToShortNumber()
         {
-			var outputsPage = new OutputsModel(new SchoolServiceMock(), httpCtxAccessor.Object);
+			var outputsPage = new OutputsModel(new MockSchoolService(), httpCtxAccessor.Object);
 			double number = 0.031416;
 			string shortNumber = outputsPage.ToShortNumber(number);
 			Assert.Equal("3.14%", shortNumber);
